PostgresTimestampWithTimeZoneTest: equality testing
Make equality testing of Date and DateComponents more rigorous.
Recognize that lossless conversion between Date and DateComponents
instances is not possible for some values (Date is implemented on
a Double). Also recognize that Date -> String rounds to millisecond
precision, and String -> Date truncates to millisecond precision
due to the implementation of DateFormatter.